refactor: move/rename designs
