--- elasticai/creator/hdl/vhdl/designs/monotonously_increasing_precomputed_scalar_function/__init__.py
@@ -1,1 +1,0 @@
-from .hard_sigmoid import HardSigmoid
