+++ elasticai/creator/hdl/vhdl/designs/monotonously_increasing_precomputed_scalar_function/design.tpl.vhd
