+++ elasticai/creator/nn/vhdl/fp_hard_sigmoid/fp_hard_sigmoid.py
@@ -4,10 +4,8 @@ from typing import Protocol
 from elasticai.creator.base_modules.hard_sigmoid import HardSigmoid as _HardSigmoidLayer
 from elasticai.creator.hdl.design_base.design import Design
 from elasticai.creator.hdl.translatable import Translatable
-from elasticai.creator.hdl.vhdl.designs.monotonously_increasing_precomputed_scalar_function.hard_sigmoid import (
-    HardSigmoid,
-)
 from elasticai.creator.nn.two_complement_fixed_point_config import FixedPointConfig
+from elasticai.creator.nn.vhdl.fp_hard_sigmoid.design import HardSigmoid
 
 
 class FixedPointOps(Protocol):
