+++ elasticai/creator/nn/vhdl/sequential/sequential.py
@@ -5,9 +5,7 @@ import torch
 
 from elasticai.creator.hdl.design_base.design import Design
 from elasticai.creator.hdl.translatable import Translatable
-from elasticai.creator.hdl.vhdl.designs.sequential import (
-    Sequential as _SequentialDesign,
-)
+from elasticai.creator.nn.vhdl.sequential.design import Sequential as _SequentialDesign
 
 
 class Sequential(Translatable, torch.nn.Sequential):
